pyproject.toml
[project]
name = "genie-space-sp-demo-v2"
version = "0.2.0"
description = "Genie Space SP demo v2 — LangGraph supervisor + Lakebase memory"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115.6",
  "uvicorn[standard]>=0.34.0",
  "httpx>=0.28.1",
  "itsdangerous>=2.2.0",
  "python-dotenv>=1.0.1",
  "pydantic>=2.10.4",

  # Agent + memory layer
  "databricks-sdk>=0.65.0",
  "databricks-langchain>=0.19.0",
  "langgraph-checkpoint-postgres>=2.0.10",
  "psycopg[binary,pool]>=3.2.3",
  "tiktoken>=0.8.0",
  "mlflow>=3.11.1",
]

[tool.uv]
# Package mode off: this is an application, not a library.
package = false
